+++ flask_unchained/bundles/sqlalchemy/alembic/migrations.py
@@ -9,4 +9,9 @@ def render_migration_item(type_, obj, autogen_context):
         autogen_context.imports.add(f'import {sqla_bundle} as sqla_bundle')
         return f'sqla_bundle.{repr(obj)}'
 
+    elif not (obj.__module__.startswith('sqlalchemy')
+              or obj.__module__.startswith('flask_unchained')):
+        autogen_context.imports.add(f'import {obj.__module__}')
+        return f'{obj.__module__}.{repr(obj)}'
+
     return False