+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## v0.9.1 (unreleased)
 
 - do not require the API bundle to be enabled to use the Security Bundle
+- fix auto-generated imports in db migrations requiring user app code
 
 ## v0.9.0 (2021/06/07)